Update audioPlayer.py
Fixes the issue where it would sometimes play the first chunk twice when the playback loop starts (i.e.. beginning of a track or after pause)

Learning: Don't call the i2s_callback directly from anywhere. Call PlayChunk instead, and it will cause the callback to be called and the playback loop to initiate. Don't know why this works differently, but it does so don't mess with it.